expose useInstance in context
