+++ src/frontend/lib/types/Rendering.tsx
@@ -1,4 +1,10 @@
+import type {
+  QueryObserverResult,
+  UseQueryResult
+} from '@tanstack/react-query';
+import type { ApiEndpoints } from '..';
 import type { ModelType } from '../enums/ModelType';
+import type { PathParams } from './Core';
 
 /**
  * Interface for rendering a model instance.
@@ -22,4 +28,27 @@ export type ModelRendererDict = EnumDictionary<
 
 export type RenderInstanceProps = {
   model: ModelType | undefined;
+  custom_model?: string;
 } & InstanceRenderInterface;
+
+export interface UseInstanceResult {
+  instance: any;
+  setInstance: (instance: any) => void;
+  refreshInstance: () => void;
+  refreshInstancePromise: () => Promise<QueryObserverResult<any, any>>;
+  instanceQuery: UseQueryResult;
+  isLoaded: boolean;
+}
+
+export interface useInstanceProps {
+  endpoint: ApiEndpoints;
+  pk?: string | number | undefined;
+  hasPrimaryKey?: boolean;
+  params?: any;
+  pathParams?: PathParams;
+  disabled?: boolean;
+  defaultValue?: any;
+  refetchOnMount?: boolean;
+  refetchOnWindowFocus?: boolean;
+  updateInterval?: number;
+}
